+++ src/index.ts
@@ -1,16 +1,39 @@
-import type { MiddlewareHandler } from 'hono';
-import { createMiddleware } from 'hono/factory';
+/**
+ * @module
+ * Event Emitter Middleware for Hono.
+ */
+
+import type { Context, Env, MiddlewareHandler } from 'hono';
 
 export type EventKey = string | symbol;
-export type EventHandler<T> = (payload: T) => void;
+export type EventHandler<T, E extends Env = Env> = (c: Context<E>, payload: T) => void | Promise<void>;
 export type EventHandlers<T> = { [K in keyof T]?: EventHandler<T[K]>[] };
 
-
 export interface Emitter<EventHandlerPayloads> {
-  on<Key extends keyof EventHandlerPayloads>(type: Key, handler: EventHandler<EventHandlerPayloads[Key]>): void;
-  off<Key extends keyof EventHandlerPayloads>(type: Key, handler?: EventHandler<EventHandlerPayloads[Key]>): void;
-  emit<Key extends keyof EventHandlerPayloads>(type: Key, payload: EventHandlerPayloads[Key]): void;
+  on<Key extends keyof EventHandlerPayloads>(key: Key, handler: EventHandler<EventHandlerPayloads[Key]>): void;
+  off<Key extends keyof EventHandlerPayloads>(key: Key, handler?: EventHandler<EventHandlerPayloads[Key]>): void;
+  emit<Key extends keyof EventHandlerPayloads>(key: Key, c: Context, payload: EventHandlerPayloads[Key]): void;
 }
+
+/**
+ * Function to define fully typed event handler.
+ * @param {EventHandler} handler - The event handlers.
+ * @returns The event handler.
+ */
+export const defineHandler = <T, K extends keyof T, E extends Env = Env>(
+  handler: EventHandler<T[K], E>,
+): EventHandler<T[K], E> => {
+  return handler;
+};
+
+/**
+ * Function to define fully typed event handlers.
+ * @param {EventHandler[]} handlers - An object where each key is an event type and the value is an array of event handlers.
+ * @returns The event handlers.
+ */
+export const defineHandlers = <T, E extends Env = Env>(handlers: { [K in keyof T]?: EventHandler<T[K], E>[] }) => {
+  return handlers;
+};
 
 /**
  * Create Event Emitter instance.
@@ -19,32 +42,54 @@ export interface Emitter<EventHandlerPayloads> {
  * @returns {Emitter} The EventEmitter instance.
  *
  * @example
- * ```ts
- * type AvailableEvents = {
- *   // event key: payload type
- *   'foo': number;
- *   'bar': { item: { id: string }, c: Context };
- * };
- *
+ * ```js
  * // Define event handlers
- * const handlers: EventHandlers<AvailableEvents> = {
+ * const handlers: {
  *   'foo': [
- *     (payload) => { console.log('Foo:', payload) }  // payload will be inferred as number
+ *     (c, payload) => { console.log('Foo:', payload) }
  *   ]
  * }
  *
  * // Initialize emitter with handlers
- * const ee = createEmitter<AvailableEvents>(handlers)
+ * const ee = createEmitter(handlers)
  *
  * // AND/OR add more listeners on the fly.
- * ee.on('bar', ({ item, c }) => {
- *   c.get('logger').log('Bar:', item.id)
+ * ee.on('bar', (c, payload) => {
+ *   c.get('logger').log('Bar:', payload.item.id)
  * })
  *
  * // Use the emitter to emit events.
- * ee.emit('foo', 42) // Payload will be expected to be of a type number
- * ee.emit('bar', { item: { id: '12345678' }, c }) // Payload will be expected to be of a type { item: { id: string }, c: Context }
+ * ee.emit('foo', c, 42)
+ * ee.emit('bar', c, { item: { id: '12345678' } })
  * ```
+ *
+ * ```ts
+ * type AvailableEvents = {
+ *   // event key: payload type
+ *   'foo': number;
+ *   'bar': { item: { id: string } };
+ * };
+ *
+ * // Define event handlers
+ * const handlers: defineHandlers<AvailableEvents>({
+ *   'foo': [
+ *     (c, payload) => { console.log('Foo:', payload) }  // payload will be inferred as number
+ *   ]
+ * })
+ *
+ * // Initialize emitter with handlers
+ * const ee = createEmitter(handlers)
+ *
+ * // AND/OR add more listeners on the fly.
+ * ee.on('bar', (c, payload) => {
+ *   c.get('logger').log('Bar:', payload.item.id)
+ * })
+ *
+ * // Use the emitter to emit events.
+ * ee.emit('foo', c, 42) // Payload will be expected to be of a type number
+ * ee.emit('bar', c, { item: { id: '12345678' }, c }) // Payload will be expected to be of a type { item: { id: string }, c: Context }
+ * ```
+ *
  */
 export const createEmitter = <EventHandlerPayloads>(
   eventHandlers?: EventHandlers<EventHandlerPayloads>,
@@ -74,7 +119,7 @@ export const createEmitter = <EventHandlerPayloads>(
      * Remove an event handler for the given event key.
      * If `handler` is undefined, all handlers for the given key are removed.
      * @param {string|symbol} key Type of event to unregister `handler` from
-     * @param {Function} [handler] Handler function to remove
+     * @param {Function} handler - Handler function to remove
      */
     off<Key extends keyof EventHandlerPayloads>(key: Key, handler?: EventHandler<EventHandlerPayloads[Key]>) {
       if (!handler) {
@@ -93,14 +138,15 @@ export const createEmitter = <EventHandlerPayloads>(
     /**
      * Emit an event with the given event key and payload.
      * Triggers all event handlers associated with the specified key.
-     * @param {string|symbol} key The event key
-     * @param {EventHandlerPayloads} [payload] Any value (preferably an object), passed to each invoked handler
+     * @param {string|symbol} key - The event key
+     * @param {Context} c - The current context object
+     * @param {EventHandlerPayloads[keyof EventHandlerPayloads]} payload - Data passed to each invoked handler
      */
-    emit<Key extends keyof EventHandlerPayloads>(key: Key, payload: EventHandlerPayloads[Key]) {
+    emit<Key extends keyof EventHandlerPayloads>(key: Key, c: Context, payload: EventHandlerPayloads[Key]) {
       const handlerArray = handlers.get(key as EventKey);
       if (handlerArray) {
         for (const handler of handlerArray) {
-          handler(payload);
+          handler(c, payload);
         }
       }
     },
@@ -110,44 +156,67 @@ export const createEmitter = <EventHandlerPayloads>(
 /**
  * Event Emitter Middleware for Hono.
  *
+ * @see {@link https://hono.dev/middleware/builtin/event-emitter}
+ *
  * @param {EventHandlers} eventHandlers - Event handlers to be registered.
  * @returns {MiddlewareHandler} The middleware handler function.
  *
  * @example
- * ```ts
- * type AvailableEvents = {
- *   // event key: payload type
- *   'foo': number;
- *   'bar': { item: { id: string }, c: Context };
- * };
+ * ```js
  *
  * // Define event handlers
- * const handlers: EventHandlers<AvailableEvents> = {
+ * const handlers: {
  *   'foo': [
- *     (payload) => { console.log('Foo:', payload) }  // payload will be inferred as number
+ *     (c, payload) => { console.log('Foo:', payload) }
+ *   ]
+ *   'bar': [
+ *     (c, payload) => { console.log('Bar:', payload.item.id) }
  *   ]
  * }
  *
- * const app = new Hono({ Variables: { emitter: Emitter<AvailableEvents> }})
+ * const app = new Hono()
  *
  * // Register the emitter middleware and provide it with the handlers
  * app.use('\*', emitter(handlers))
  *
- * // AND/OR create event handler as "named function"
- * const barEventHandler = ({ item, c }) => {
- *   c.get('logger').log('Bar:', item.id)
- * }
- * // and register it inside middleware or route handler
- * app.use('\*', (c, next) => {
- *   c.get('emitter').on('bar', barEventHandler)
- *   return next()
+ * // Use the emitter in route handlers to emit events.
+ * app.post('/foo', async (c) => {
+ *   // The emitter is available under "emitter" key in the context.
+ *   c.get('emitter').emit('foo', c, 42)
+ *   c.get('emitter').emit('bar', c, { item: { id: '12345678' } })
+ *   return c.text('Success')
  * })
+ * ```
+ *
+ * ```ts
+ * type AvailableEvents = {
+ *   // event key: payload type
+ *   'foo': number;
+ *   'bar': { item: { id: string } };
+ * };
+ *
+ * type Env = { Bindings: {}; Variables: { emitter: Emitter<AvailableEvents> }; }
+ *
+ * // Define event handlers
+ * const handlers: defineHandlers<AvailableEvents>({
+ *   'foo': [
+ *     (c, payload) => { console.log('Foo:', payload) }  // payload will be inferred as number
+ *   ]
+ *   'bar': [
+ *     (c, payload) => { console.log('Bar:', payload.item.id) }  // payload will be inferred as { item: { id: string } }
+ *   ]
+ * })
+ *
+ * const app = new Hono<Env>()
+ *
+ * // Register the emitter middleware and provide it with the handlers
+ * app.use('\*', emitter(handlers))
  *
  * // Use the emitter in route handlers to emit events.
  * app.post('/foo', async (c) => {
  *   // The emitter is available under "emitter" key in the context.
- *   c.get('emitter').emit('foo', 42) // Payload will be expected to be of a type number
- *   c.get('emitter').emit('bar', { item: { id: '12345678' }, c }) // Payload will be expected to be of a type { item: { id: string }, c: Context }
+ *   c.get('emitter').emit('foo', c, 42) // Payload will be expected to be of a type number
+ *   c.get('emitter').emit('bar', c, { item: { id: '12345678' } }) // Payload will be expected to be of a type { item: { id: string } }
  *   return c.text('Success')
  * })
  * ```
@@ -157,8 +226,8 @@ export const emitter = <EventHandlerPayloads>(
 ): MiddlewareHandler => {
   // Create new instance to share with any middleware and handlers
   const instance = createEmitter<EventHandlerPayloads>(eventHandlers);
-  return createMiddleware(async (c, next) => {
+  return async (c, next) => {
     c.set('emitter', instance);
     await next();
-  });
+  };
 };